# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Diom API",
    "description": "Welcome to the Diom API documentation!\n",
    "version": "0.1.0",
    "x-logo": {
      "altText": "Svix Logo",
      "url": "https://www.svix.com/static/img/brand-padded.svg"
    }
  },
  "paths": {
...
    "/api/v1.cluster-admin.remove-node": {
      "post": {
        "tags": [
          "ClusterAdmin"
        ],
        "summary": "Cluster Remove Node",
        "description": "Remove a node from the cluster.\n\nThis operation executes immediately and the node must be wiped and reset\nbefore it can safely be added to the cluster.",
        "operationId": "v1.cluster-admin.remove-node",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterRemoveNodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterRemoveNodeOut"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "HTTP Bearer token passed in the `Authorization` header"
      }
    },
    "schemas": {
...
      "ClusterRemoveNodeIn": {
        "type": "object",
        "properties": {
          "node_id": {
            "type": "string"
          }
        },
        "required": [
          "node_id"
        ]
      },
      "ClusterRemoveNodeOut": {
        "type": "object",
        "properties": {
          "node_id": {
            "type": "string"
          }
        },
        "required": [
          "node_id"
        ]
      },
...
      "StandardErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "detail"
        ]
      },
...
      "ValidationErrorBody": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            }
          }
        },
        "required": [
          "detail"
        ]
      },
      "ValidationErrorItem": {
        "description": "Validation errors have their own schema to provide context for invalid requests eg. mismatched\ntypes and out of bounds values. There may be any number of these per 422 UNPROCESSABLE ENTITY\nerror.",
        "type": "object",
        "properties": {
          "loc": {
            "description": "The location as a [`Vec`] of [`String`]s -- often in the form `[\"body\", \"field_name\"]`,\n`[\"query\", \"field_name\"]`, etc. They may, however, be arbitrarily deep.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "description": "The message accompanying the validation error item.",
            "type": "string"
          },
          "type": {
            "description": "The type of error.\n\nOften \"type_error\" or \"value_error\", but sometimes with more context like\n\"value_error.number.not_ge\".",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ]
      }
    }
  },
...
}